+++ src/common/modals/Settings/components/General.js
@@ -5,34 +5,20 @@ import { useSelector, useDispatch } from 'react-redux';
 import path from 'path';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import fsa from 'fs-extra';
-import { promises as fs } from 'fs';
 import {
   faCopy,
   faDownload,
   faTachometerAlt,
-  faTrash,
-  faPlay,
   faToilet,
-  faNewspaper,
-  faFolder,
   faFire
 } from '@fortawesome/free-solid-svg-icons';
-import { Select, Tooltip, Button, Switch, Input, Checkbox } from 'antd';
-import { faDiscord } from '@fortawesome/free-brands-svg-icons';
+import { Select, Tooltip, Button, Switch } from 'antd';
+import { _getCurrentAccount } from '../../../utils/selectors';
 import {
-  _getCurrentAccount,
-  _getDataStorePath,
-  _getInstancesPath,
-  _getTempPath
-} from '../../../utils/selectors';
-import {
-  updateDiscordRPC,
-  updateHideWindowOnGameLaunch,
   updatePotatoPcMode,
-  updateShowNews,
   updateCurseReleaseChannel
 } from '../../../reducers/settings/actions';
-import HorizontalLogo from '../../../../ui/HorizontalLogo';
+import HorizontalLogo from '../../../../ui/HorizontalLogo.png';
 import { updateConcurrentDownloads } from '../../../reducers/actions';
 import { openModal } from '../../../reducers/modals/actions';
 import { extractFace } from '../../../../app/desktop/utils';
@@ -65,7 +51,7 @@ const ProfileImage = styled.img`
   position: relative;
   top: 20px;
   left: 20px;
-  background: #212b36;
+  background: #3d3d3d;
   width: 50px;
   height: 50px;
 `;
@@ -74,7 +60,7 @@ const ImagePlaceHolder = styled.div`
   position: relative;
   top: 20px;
   left: 20px;
-  background: #212b36;
+  background: #3d3d3d;
   width: 50px;
   height: 50px;
 `;
@@ -149,7 +135,7 @@ const ParallelDownload = styled.div`
   }
 `;
 
-const DiscordRpc = styled.div`
+const ListView = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: space-between;
@@ -171,22 +157,6 @@ const LauncherVersion = styled.div`
 
   h1 {
     color: ${props => props.theme.palette.text.primary};
-  }
-`;
-const CustomDataPathContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 140px;
-  border-radius: ${props => props.theme.shape.borderRadius};
-
-  h1 {
-    width: 100%;
-    font-size: 15px;
-    font-weight: 700;
-    color: ${props => props.theme.palette.text.primary};
-    z-index: 1;
-    text-align: left;
   }
 `;
 
@@ -213,37 +183,19 @@ const General = () => {
   const [version, setVersion] = useState(null);
   const [releaseChannel, setReleaseChannel] = useState(null);
   const currentAccount = useSelector(_getCurrentAccount);
-  const hideWindowOnGameLaunch = useSelector(
-    state => state.settings.hideWindowOnGameLaunch
-  );
-  const DiscordRPC = useSelector(state => state.settings.discordRPC);
+
   const potatoPcMode = useSelector(state => state.settings.potatoPcMode);
   const concurrentDownloads = useSelector(
     state => state.settings.concurrentDownloads
   );
   const updateAvailable = useSelector(state => state.updateAvailable);
-  const dataStorePath = useSelector(_getDataStorePath);
-  const instancesPath = useSelector(_getInstancesPath);
-  const isPlaying = useSelector(state => state.startedInstances);
-  const queuedInstances = useSelector(state => state.downloadQueue);
-  const tempPath = useSelector(_getTempPath);
   const [copiedUuid, setCopiedUuid] = useState(false);
   const [profileImage, setProfileImage] = useState(null);
-  const [deletingInstances, setDeletingInstances] = useState(false);
-  const userData = useSelector(state => state.userData);
-  const [dataPath, setDataPath] = useState(userData);
-  const [moveUserData, setMoveUserData] = useState(false);
-  const showNews = useSelector(state => state.settings.showNews);
-  const [loadingMoveUserData, setLoadingMoveUserData] = useState(false);
   const curseReleaseChannel = useSelector(
     state => state.settings.curseReleaseChannel
   );
 
   const dispatch = useDispatch();
-
-  const disableInstancesActions =
-    Object.keys(queuedInstances).length > 0 ||
-    Object.keys(isPlaying).length > 0;
 
   useEffect(() => {
     ipcRenderer.invoke('getAppVersion').then(setVersion).catch(console.error);
@@ -258,66 +210,6 @@ const General = () => {
       )
       .catch(console.error);
   }, []);
-
-  const clearSharedData = async () => {
-    setDeletingInstances(true);
-    try {
-      await fsa.emptyDir(dataStorePath);
-      await fsa.emptyDir(instancesPath);
-      await fsa.emptyDir(tempPath);
-    } catch (e) {
-      console.error(e);
-    }
-    setDeletingInstances(false);
-  };
-
-  const changeDataPath = async () => {
-    setLoadingMoveUserData(true);
-    const appData = await ipcRenderer.invoke('getAppdataPath');
-    const appDataPath = path.join(appData, 'koalalauncher');
-
-    const notCopiedFiles = [
-      'Cache',
-      'Code Cache',
-      'Dictionaries',
-      'GPUCache',
-      'Cookies',
-      'Cookies-journal'
-    ];
-    await fsa.writeFile(path.join(appDataPath, 'override.data'), dataPath);
-
-    if (moveUserData) {
-      try {
-        const files = await fs.readdir(userData);
-        await Promise.all(
-          files.map(async name => {
-            if (!notCopiedFiles.includes(name)) {
-              await fsa.copy(
-                path.join(userData, name),
-                path.join(dataPath, name),
-                {
-                  overwrite: true
-                }
-              );
-            }
-          })
-        );
-      } catch (e) {
-        console.error(e);
-      }
-    }
-    setLoadingMoveUserData(false);
-    ipcRenderer.invoke('appRestart');
-  };
-
-  const openFolder = async () => {
-    const { filePaths, canceled } = await ipcRenderer.invoke(
-      'openFolderDialog',
-      userData
-    );
-    if (!filePaths[0] || canceled) return;
-    setDataPath(filePaths[0]);
-  };
 
   return (
     <MyAccountPrf>
@@ -375,7 +267,7 @@ const General = () => {
         <div>
           <div
             css={`
-              width: 400px;
+              width: 450px;
             `}
           >
             Stable updates once a month, beta does update more often but it may
@@ -400,15 +292,18 @@ const General = () => {
           </Select>
         </div>
       </ReleaseChannel>
-      <Hr />
-      <Title>
+      <Title
+        css={`
+          margin-top: 5px;
+        `}
+      >
         Concurrent Downloads &nbsp; <FontAwesomeIcon icon={faTachometerAlt} />
       </Title>
       <ParallelDownload>
         <p
           css={`
             margin: 0;
-            width: 400px;
+            width: 450px;
           `}
         >
           Select the number of concurrent downloads. If you have a slow
@@ -423,7 +318,7 @@ const General = () => {
             text-align: start;
           `}
         >
-          {[...Array(20).keys()]
+          {[...Array(100).keys()]
             .map(x => x + 1)
             .map(x => (
               <Select.Option key={x} value={x}>
@@ -434,17 +329,18 @@ const General = () => {
       </ParallelDownload>
       <Hr />
       <Title>
-        Preferred Curse Release Channel &nbsp; <FontAwesomeIcon icon={faFire} />
+        Preferred CurseForge Release Channel &nbsp;{' '}
+        <FontAwesomeIcon icon={faFire} />
       </Title>
       <ParallelDownload>
         <p
           css={`
             margin: 0;
-            width: 400px;
+            width: 450px;
           `}
         >
-          Select the preferred release channel for downloading Curse projects.
-          This also applies for mods update.
+          Select the preferred release channel for downloading CurseForge
+          projects. This also applies for mods update.
         </p>
         <Select
           css={`
@@ -462,94 +358,14 @@ const General = () => {
       <Hr />
       <Title
         css={`
-          margin-top: 0px;
-        `}
-      >
-        Discord Integration &nbsp; <FontAwesomeIcon icon={faDiscord} />
-      </Title>
-      <DiscordRpc>
-        <p
-          css={`
-            width: 350px;
-          `}
-        >
-          Enable / disable Discord Integration. This displays what you are
-          playing in Discord.
-        </p>
-        <Switch
-          onChange={e => {
-            dispatch(updateDiscordRPC(e));
-            if (e) {
-              ipcRenderer.invoke('init-discord-rpc');
-            } else {
-              ipcRenderer.invoke('shutdown-discord-rpc');
-            }
-          }}
-          checked={DiscordRPC}
-        />
-      </DiscordRpc>
-      <Hr />
-      <Title
-        css={`
-          margin-top: 0px;
-        `}
-      >
-        Minecraft News &nbsp; <FontAwesomeIcon icon={faNewspaper} />
-      </Title>
-      <DiscordRpc>
-        <p
-          css={`
-            width: 350px;
-          `}
-        >
-          Enable / disable Minecraft news.
-        </p>
-        <Switch
-          onChange={e => {
-            dispatch(updateShowNews(e));
-          }}
-          checked={showNews}
-        />
-      </DiscordRpc>
-      <Hr />
-      <Title
-        css={`
-          margin-top: 0px;
-        `}
-      >
-        Hide Launcher While Playing &nbsp; <FontAwesomeIcon icon={faPlay} />
-      </Title>
-      <DiscordRpc
-        css={`
-          margin-bottom: 30px;
-        `}
-      >
-        <p
-          css={`
-            width: 500px;
-          `}
-        >
-          Automatically hide the launcher when launching an instance. You will
-          still be able to open it from the icon tray
-        </p>
-        <Switch
-          onChange={e => {
-            dispatch(updateHideWindowOnGameLaunch(e));
-          }}
-          checked={hideWindowOnGameLaunch}
-        />
-      </DiscordRpc>
-      <Hr />
-      <Title
-        css={`
-          margin-top: 0px;
+          margin-top: 5px;
         `}
       >
         Potato PC Mode &nbsp; <FontAwesomeIcon icon={faToilet} />
       </Title>
-      <DiscordRpc
+      <ListView
         css={`
-          margin-bottom: 30px;
+          margin-bottom: 10px;
         `}
       >
         <p
@@ -566,143 +382,7 @@ const General = () => {
           }}
           checked={potatoPcMode}
         />
-      </DiscordRpc>
-      <Hr />
-      <Title
-        css={`
-          width: 300px;
-          float: left;
-        `}
-      >
-        Clear Shared Data&nbsp; <FontAwesomeIcon icon={faTrash} />
-      </Title>
-      <div
-        css={`
-          display: flex;
-          justify-content: space-between;
-          text-align: left;
-          width: 100%;
-          margin-bottom: 30px;
-          p {
-            text-align: left;
-            color: ${props => props.theme.palette.text.third};
-          }
-        `}
-      >
-        <p
-          css={`
-            margin: 0;
-            width: 500px;
-          `}
-        >
-          Deletes all the shared files between instances. Doing this will result
-          in the complete loss of the instances data
-        </p>
-        <Button
-          onClick={() => {
-            dispatch(
-              openModal('ActionConfirmation', {
-                message: 'Are you sure you want to delete shared data?',
-                confirmCallback: clearSharedData,
-                title: 'Confirm'
-              })
-            );
-          }}
-          disabled={disableInstancesActions}
-          loading={deletingInstances}
-        >
-          Clear
-        </Button>
-      </div>
-      <Hr />
-      {/* {process.env.REACT_APP_RELEASE_TYPE === 'setup' && ( */}
-      <CustomDataPathContainer>
-        <Title
-          css={`
-            width: 400px;
-            float: left;
-          `}
-        >
-          User Data Path&nbsp; <FontAwesomeIcon icon={faFolder} />
-          <a
-            css={`
-              margin-left: 30px;
-            `}
-            onClick={async () => {
-              const appData = await ipcRenderer.invoke('getAppdataPath');
-              const appDataPath = path.join(appData, 'koalalauncher');
-              setDataPath(appDataPath);
-            }}
-          >
-            Reset Path
-          </a>
-        </Title>
-        <div
-          css={`
-            display: flex;
-            justify-content: space-between;
-            text-align: left;
-            width: 100%;
-            height: 30px;
-            margin: 20px 0 10px 0;
-            p {
-              text-align: left;
-              color: ${props => props.theme.palette.text.third};
-            }
-          `}
-        >
-          <Input
-            value={dataPath}
-            onChange={e => setDataPath(e.target.value)}
-            disabled={
-              loadingMoveUserData ||
-              deletingInstances ||
-              disableInstancesActions
-            }
-          />
-          <Button
-            css={`
-              margin-left: 20px;
-            `}
-            onClick={openFolder}
-            disabled={loadingMoveUserData || deletingInstances}
-          >
-            <FontAwesomeIcon icon={faFolder} />
-          </Button>
-          <Button
-            css={`
-              margin-left: 20px;
-            `}
-            onClick={changeDataPath}
-            disabled={
-              disableInstancesActions ||
-              userData === dataPath ||
-              !dataPath ||
-              dataPath.length === 0 ||
-              deletingInstances
-            }
-            loading={loadingMoveUserData}
-          >
-            Apply & Restart
-          </Button>
-        </div>
-        <div
-          css={`
-            display: flex;
-            justify-content: flex-start;
-            width: 100%;
-          `}
-        >
-          <Checkbox
-            onChange={e => {
-              setMoveUserData(e.target.checked);
-            }}
-          >
-            Copy current data to the new directory
-          </Checkbox>
-        </div>
-      </CustomDataPathContainer>
-      {/* )} */}
+      </ListView>
       <Hr />
       <LauncherVersion>
         <div
@@ -713,17 +393,27 @@ const General = () => {
             margin: 10px 0;
           `}
         >
-          <HorizontalLogo
-            size={200}
-            onClick={() => dispatch(openModal('ChangeLogs'))}
+          {/* eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions */}
+          <img
+            src={HorizontalLogo}
+            height="128px"
+            width="311px"
+            alt="Logo"
+            draggable="false"
+            onClick={() => dispatch(openModal('Changelogs'))}
+            pointerCursor
           />{' '}
-          <div
-            css={`
-              margin-left: 10px;
-            `}
-          >
-            v {version}
-          </div>
+        </div>
+        <div
+          css={`
+            margin-top: -40px;
+            margin-left: 230px;
+            margin-bottom: 50px;
+            color: #6c5d53;
+            font-size: 18px;
+          `}
+        >
+          <p>v{version}</p>
         </div>
         <p>
           {updateAvailable
@@ -732,7 +422,7 @@ const General = () => {
         </p>
         <div
           css={`
-            margin-top: 20px;
+            margin-top: 5px;
             height: 36px;
             display: flex;
             flex-direction: row;
