+++ src/common/modals/Settings/components/Graphics.js
@@ -1,0 +1,145 @@
+import React, { useState, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import styled from 'styled-components';
+import { ipcRenderer } from 'electron';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faDesktop } from '@fortawesome/free-solid-svg-icons';
+import { Input, Select } from 'antd';
+import { updateResolution } from '../../../reducers/settings/actions';
+import { resolutionPresets } from '../../../../app/desktop/utils/constants';
+
+const Graphics = styled.div`
+  width: 100%;
+  height: 400px;
+`;
+
+const Resolution = styled.div`
+  width: 100%;
+  height: 100px;
+`;
+
+const ResolutionInputContainer = styled.div`
+  margin-top: 10px;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+
+  div {
+    width: 200px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+  }
+`;
+
+const Title = styled.h3`
+  font-size: 15px;
+  font-weight: 700;
+  color: ${props => props.theme.palette.text.secondary};
+`;
+
+const Paragraph = styled.p`
+  color: ${props => props.theme.palette.text.third};
+`;
+
+const Hr = styled.div`
+  height: 35px;
+`;
+
+const MainTitle = styled.h1`
+  color: ${props => props.theme.palette.text.primary};
+  width: 80px;
+  margin: 30px 0 20px 0;
+`;
+
+export default function MyAccountPreferences() {
+  const [screenResolution, setScreenResolution] = useState(null);
+  const mcResolution = useSelector(
+    state => state.settings.minecraftSettings.resolution
+  );
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    ipcRenderer
+      .invoke('getAllDisplaysBounds')
+      .then(setScreenResolution)
+      .catch(console.error);
+  }, []);
+
+  return (
+    <Graphics>
+      <MainTitle>Graphics</MainTitle>
+      <Resolution>
+        <Title
+          css={`
+            width: 100%;
+            margin-top: 0px;
+            height: 8px;
+            text-align: left;
+            margin-bottom: 20px;
+          `}
+        >
+          Game Resolution&nbsp; <FontAwesomeIcon icon={faDesktop} />
+        </Title>
+        <Paragraph
+          css={`
+            width: 100%;
+            text-align: left;
+            margin: 0;
+          `}
+        >
+          Select the initial game resolution in pixels (width x height)
+        </Paragraph>
+        <ResolutionInputContainer>
+          <div>
+            <Input
+              placeholder="width"
+              value={mcResolution.width}
+              onChange={e => {
+                const w = parseInt(e.target.value, 10);
+                dispatch(updateResolution({ width: w || 854 }));
+              }}
+            />
+            &nbsp;X&nbsp;
+            <Input
+              placeholder="Height"
+              value={mcResolution.height}
+              onChange={e => {
+                const h = parseInt(e.target.value, 10);
+                dispatch(updateResolution({ height: h || 480 }));
+              }}
+            />
+          </div>
+          <Select
+            placeholder="Presets"
+            onChange={v => {
+              const w = parseInt(v.split('x')[0], 10);
+              const h = parseInt(v.split('x')[1], 10);
+              dispatch(updateResolution({ height: h, width: w }));
+            }}
+          >
+            {resolutionPresets.map(v => {
+              const w = parseInt(v.split('x')[0], 10);
+              const h = parseInt(v.split('x')[1], 10);
+
+              const isBiggerThanScreen = (screenResolution || []).every(
+                bounds => {
+                  return bounds.width < w || bounds.height < h;
+                }
+              );
+              if (isBiggerThanScreen) return null;
+              return (
+                <Select.Option key={v} value={v}>
+                  {v}
+                </Select.Option>
+              );
+            })}
+          </Select>
+        </ResolutionInputContainer>
+      </Resolution>
+      <Hr />
+    </Graphics>
+  );
+}
